refactor: MA-28 pressentation 레이어 패키지 정리
