+++ app/src/main/java/com/mundcode/muntam/presentation/screen/exams/ExamsScreen.kt
@@ -1,4 +1,4 @@
-package com.mundcode.muntam.presentation.ui.exams
+package com.mundcode.muntam.presentation.screen.exams
 
 import android.Manifest
 import android.util.Log
