+++ app/src/main/java/com/mundcode/muntam/presentation/screen/subject_add/SubjectAddViewModel.kt
@@ -1,4 +1,4 @@
-package com.mundcode.muntam.presentation.ui.subject_add
+package com.mundcode.muntam.presentation.screen.subject_add
 
 import com.mundcode.muntam.base.BaseViewModel
