+++ app/src/main/java/com/mundcode/muntam/presentation/screen/subject_setting/SubjectSettingsScreen.kt
@@ -1,4 +1,4 @@
-package com.mundcode.muntam.presentation.ui.subject_setting
+package com.mundcode.muntam.presentation.screen.subject_setting
 
 import androidx.compose.runtime.Composable
 import com.mundcode.muntam.Dummy
